requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==24.2.0
basicsr==1.4.2
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
chardet==4.0.0
charset-normalizer==3.4.3
click==8.1.8
cycler==0.11.0
dataclasses==0.6
DateTime==5.5
discord==1.7.3
discord.py==2.6.3
facexlib==0.3.0
ffmpeg==1.4
filterpy==1.4.5
Flask==2.2.5
fonttools==4.38.0
future==1.0.0
gfpgan==1.3.8
google-auth==2.40.3
google-auth-oauthlib==0.4.6
grpcio==1.62.3
idna==3.10
imageio==2.31.2
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.6
kiwisolver==1.4.5
llvmlite==0.39.1
lmdb==1.7.3
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
multidict==6.0.5
mutagen==1.47.0
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.12.0.88
packaging==24.0
Pillow==9.5.0
platformdirs==4.0.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodomex==3.23.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyWavelets==1.3.0
PyYAML==6.0.1
realesrgan==0.3.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-image==0.19.3
scipy==1.7.3
six==1.17.0
tb-nightly==2.12.0a20230113
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2021.11.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.67.1
typing_extensions==4.7.1
urllib3==2.0.7
websockets==13.1
Werkzeug==2.2.3
yapf==0.43.0
yarl==1.9.4
yt-dlp==2024.10.22
zipp==3.15.0
zope.interface==6.4.post2
